SAS-07: refactored landing and code cleanup
