--- libs/pages/src/constants/index.ts
@@ -1,13 +1,0 @@
-export const h1Content =
-  'Welcome to Stone & Sable, where functionality meets artistry in every ' +
-  'piece we offer. Illuminate your space with our curated collection of ' +
-  'distinctive lamps, anchor your rooms with our quality furniture, and ' +
-  'find the perfect seating that invites both comfort and conversation.'
-
-export const highlightWords = [
-  'functionality',
-  'artistry',
-  'quality',
-  'comfort',
-  'conversation.',
-]
