+++ libs/shared/ui/src/base/index.ts
@@ -1,0 +1,7 @@
+export * from './button'
+export * from './dropDownMenu'
+export * from './fadeWordsIn'
+export * from './smoothScroller'
+export * from './sofaBackDrop'
+export * from './seperator'
+export * from './card'
