+++ libs/shared/ui/src/composite/index.ts
@@ -1,0 +1,7 @@
+export * from './header/UserAuthHeader'
+export * from './navigation/NavBar'
+export * from './navigation/NavLinks'
+export * from './products/ProductsList'
+export * from './products/FeaturedProducts'
+export * from './landing/HeroProducts'
+export * from './landing/HeroDescriptor'
