+++ libs/shared/ui/src/index.ts
@@ -1,11 +1,2 @@
-export * from './base/button'
-export * from './base/dropDownMenu'
-export * from './base/fadeWordsIn'
-export * from './base/smoothScroller'
-export * from './base/sofaBackDrop'
-export * from './base/seperator'
-export * from './base/card'
-
-export * from './composite/header/UserAuthHeader'
-export * from './composite/navigation/NavBar'
-export * from './composite/navigation/NavLinks'
+export * from './base'
+export * from './composite'
